Conflicts:
	Libs/include/libpvr/PVRTGlobal.h
	Projects/MacOSFramework/Framework.xcodeproj/project.pbxproj
	Projects/UnitTests/Classes/GameCore.cpp
	Projects/UnitTests/TemplateProjectMacOS.xcodeproj/project.pbxproj
	Projects/UnitTests/TemplateProjectiPhone.xcodeproj/project.pbxproj
	Projects/Win32Framework/Win32Framework.vcxproj.filters
	Projects/iPhoneFramework/Framework.xcodeproj/project.pbxproj
	Sources/Internal/Render/Image.cpp
	Sources/Internal/Render/ImageLoader.cpp
	Sources/Internal/Render/LibDdsHelper.cpp
	Sources/Internal/Render/LibDdsHelper.h
	Sources/Internal/Render/LibPVRHelper.cpp
	Sources/Internal/Render/LibPVRHelper.h
	Sources/Internal/Render/LibPngHelpers.cpp
	Sources/Internal/Render/Material/NMaterial.cpp
	Sources/Internal/Render/Texture.cpp
	Sources/Tools/TextureCompression/DXTConverter.cpp
	Tools/ResourceEditor/Classes/ImageTools/ImageTools.cpp
	Tools/ResourceEditor/ResourceEditorQt.xcodeproj/project.pbxproj
	Tools/ResourceEditor/ResourceEditorQtVS2010.vcxproj.filters

+++ Tools/ResourceEditor/Classes/Qt/Scene/System/LandscapeEditorDrawSystem/HeightmapProxy.cpp
@@ -39,7 +39,7 @@ void HeightmapProxy::UpdateRect(const DAVA::Rect &rect)
 {
 	int32 size = Size();
 
-	Rect bounds(0.f, 0.f, size - 1, size - 1);
+	Rect bounds(0.f, 0.f, (float32)(size - 1), (float32)(size - 1));
 	changedRect = rect;
 	bounds.ClampToRect(changedRect);
